+++ cmd/lino/dep_push.go
@@ -26,7 +26,9 @@ import (
 
 func pushDataDestinationFactory() map[string]domain.DataDestinationFactory {
 	return map[string]domain.DataDestinationFactory{
-		"postgres": infra.NewPostgresDataDestinationFactory(logger),
+		"postgres":   infra.NewPostgresDataDestinationFactory(logger),
+		"godror":     infra.NewOracleDataDestinationFactory(logger),
+		"godror-raw": infra.NewOracleDataDestinationFactory(logger),
 	}
 }